Use Makefile to generate files
If we make the add_nerc_rates.py script generic, so that it accepts
filenames on the command line rather than hardcoding them, we get a
solution that permits us to add templating to any file simply by renaming
it from `foo.md` to `foo.in.md`.

We update the `mkdocs` configuration so that it excludes `*.in.md` files
from the published documentation.